Pull some classes out of dumping ground files.
